ichannel_rust_tests::bindings

Function connectx_as

pub unsafe extern "C" fn connectx_as(
    sock: i32,
    addrs: *const sockunion_t,
    cnt: i32,
    src: *const sockunion_t,
    type_: i32,
    proto: i32,
    flags: i32,
    timeout: i32,
) -> i32

Connect using a specified src

\param[in] sock a file descriptor for the socket, a negative value to create a new socket \param[in] addrs addresses to connect to \param[in] cnt number of addresses to connect to (=1 except for SCTP) \param[in] src use specific network interface as source \param[in] type see socket(2) \param[in] proto see socket(2) \param[in] flags see fd_features_flags_t \param[in] timeout the maximum time, in seconds, spent to establish the connection; only for blocking connects; bound to the socket and kept after connection such that further blocking read attempts would also have this timeout; ignored if equals to 0.

\Returns On success, the file descriptor for the socket On error, -1 and errno is set appropriately.