ichannel_rust_tests::bindings

Function module_check_no_dependencies

pub unsafe extern "C" fn module_check_no_dependencies(
    tab: *mut *mut module_t,
    len: i32,
    collision: *mut lstr_t,
) -> i32

Checks if the given modules are independent from each other.

\param[in] tab input modules list. \param[in] len length of the modules list. \param[out] collision name of the first module in the list on which another module from the list depends (if any).

\return -1 in case of collision.