ichannel_rust_tests::bindings

Function ps_split_escaped

pub unsafe extern "C" fn ps_split_escaped(
    mp: *mut mem_pool_t,
    ps: pstream_t,
    sep: *const ctype_desc_t,
    escape: i8,
    flags: u32,
    res: *mut qv_lstr_t,
)

Split a stream based on a set of separator and an escape character.

The line is parsed and

Each time one of the separators is encountered, a new chunk is added in the result vector.

Each time the escape character is encountered:

If the next character is a separator or an escape character. This character is added in the result vector.

If not, the escape character and the next character are added in the result vector.

Strings appended are copied. The results may contain empty strings if flags == PS_SPLIT_SKIP_EMPTY.

\param[in] mp The pool on which copied data is allocated (NULL to allocate data on the heap). \param[in] ps The input stream. \param[in] sep The separating characters. \param[in] esc An escape character. \param[in] flags Some flags (see the enum declaration above) \param[out] res A vector that get filled with the content of the ps.