ichannel_rust_tests::bindings

Function sb_add_compressed

pub unsafe extern "C" fn sb_add_compressed(
    out: *mut sb_t,
    data: *const c_void,
    dlen: usize,
    level: i32,
    do_gzip: bool,
) -> isize

Add compressed data in the string buffer.

Takes the chunk of data pointed by @p data and @p dlen and add it compressed using zlib in the sb_t @p out.

@param out output buffer @param data source data @param dlen size of the data pointed by @p data. @param level compression level @param do_gzip if true compresses using ‘gzip’, else use ‘deflate’

@return amount of data written in the stream or a negative zlib error in case of error.