ichannel_rust_tests::bindings

Function ssl_do_handshake

pub unsafe extern "C" fn ssl_do_handshake(
    ssl: *mut ssl_st,
    ev: *mut ev_t,
    fd: i32,
    rbuf: *mut sb_t,
) -> ssl_handshake_status_t

Wrapper to SSL_do_handshake.

\param[in] ssl The ssl context on which the handshake is done. \param[in] ev The fd el_t used to change its mask (POLLIN, POLLOUT). Nullable in the case of a blocking socket. \param[in] fd The fd \param[in,out] rbuf If set, this buffer is considered to have the first octets of the handshake. When it becomes empty, data are read from fd and the buffer is reset. \return the handshake’s status